+++ lambdas/data-migration1/tests/test-async-operations.js
@@ -91,7 +91,7 @@ test.serial('migrateAsyncOperationRecord correctly migrates asyncOperation recor
       ...fakeAsyncOp,
       operation_type: fakeAsyncOp.operationType,
       task_arn: fakeAsyncOp.taskArn,
-      output: JSON.parse(fakeAsyncOp.output),
+      output: { output: JSON.parse(fakeAsyncOp.output) },
       created_at: new Date(fakeAsyncOp.createdAt),
       updated_at: new Date(fakeAsyncOp.updatedAt),
     },
@@ -100,8 +100,8 @@ test.serial('migrateAsyncOperationRecord correctly migrates asyncOperation recor
 });
 
 test.serial('migrateAsyncOperationRecord correctly migrates asyncOperation record where record.output is an array', async (t) => {
-  const output = { output: '[\"string\",\"test-string"]' };
-  const fakeAsyncOp = generateFakeAsyncOperation(output);
+  const output = '[\"string\",\"test-string"]';
+  const fakeAsyncOp = generateFakeAsyncOperation({ output });
   await migrateAsyncOperationRecord(fakeAsyncOp, t.context.knex);
 
   const createdRecord = await t.context.knex.queryBuilder()
@@ -116,7 +116,7 @@ test.serial('migrateAsyncOperationRecord correctly migrates asyncOperation recor
       ...fakeAsyncOp,
       operation_type: fakeAsyncOp.operationType,
       task_arn: fakeAsyncOp.taskArn,
-      output: JSON.parse(fakeAsyncOp.output),
+      output: { output: JSON.parse(fakeAsyncOp.output) },
       created_at: new Date(fakeAsyncOp.createdAt),
       updated_at: new Date(fakeAsyncOp.updatedAt),
     },
@@ -256,4 +256,29 @@ test.serial('migrateAsyncOperations processes all non-failing records', async (t
     {}
   );
   t.is(records.length, 1);
+});
+
+test.serial('migrateAsyncOperationRecord correctly migrates asyncOperation record where record.output is a string', async (t) => {
+  const output = 'some-string';
+  const fakeAsyncOp = generateFakeAsyncOperation({ output });
+  await migrateAsyncOperationRecord(fakeAsyncOp, t.context.knex);
+
+  const createdRecord = await t.context.knex.queryBuilder()
+    .select()
+    .table('async_operations')
+    .where({ id: fakeAsyncOp.id })
+    .first();
+
+  t.deepEqual(
+    omit(createdRecord, ['cumulus_id']),
+    omit({
+      ...fakeAsyncOp,
+      operation_type: fakeAsyncOp.operationType,
+      task_arn: fakeAsyncOp.taskArn,
+      output: { output: fakeAsyncOp.output },
+      created_at: new Date(fakeAsyncOp.createdAt),
+      updated_at: new Date(fakeAsyncOp.updatedAt),
+    },
+    ['createdAt', 'updatedAt', 'operationType', 'taskArn'])
+  );
 });