+++ packages/api/lambdas/sf-scheduler.js
@@ -1,12 +1,14 @@
 'use strict';
 
 const get = require('lodash/get');
+const isNil = require('lodash/isNil');
 
+const { getCollection } = require('@cumulus/api-client/collections');
 const SQS = require('@cumulus/aws-client/SQS');
 const { getProvider } = require('@cumulus/api-client/providers');
 const { buildQueueMessageFromTemplate } = require('@cumulus/message/Build');
-const isNil = require('lodash/isNil');
-const Collection = require('../models/collections');
+const Logger = require('@cumulus/logger');
+const logger = new Logger({ sender: '@cumulus/api/lambdas/sf-scheduler' });
 
 const getApiProvider = (providerId) => {
   if (isNil(providerId)) return undefined;
@@ -16,11 +18,14 @@ const getApiProvider = (providerId) => {
   });
 };
 
-const getCollection = (collection) => {
+const getApiCollection = (collection) => {
+  logger.debug(`Getting collection from API that matches ${JSON.stringify(collection)}`);
   if (isNil(collection)) return undefined;
-
-  const c = new Collection();
-  return c.get({ name: collection.name, version: collection.version });
+  return getCollection({
+    prefix: process.env.stackName,
+    collectionName: collection.name,
+    collectionVersion: collection.version,
+  });
 };
 
 /**
@@ -35,7 +40,7 @@ const getCollection = (collection) => {
 async function handleScheduleEvent(event) {
   const [providerRecord, collection] = await Promise.all([
     getApiProvider(event.provider),
-    getCollection(event.collection),
+    getApiCollection(event.collection),
   ]);
 
   const provider = providerRecord ? JSON.parse(providerRecord.body) : undefined;