+++ packages/api/tests/endpoints/granules/test-reingest-granules.js
@@ -9,25 +9,44 @@ const {
   recursivelyDeleteS3Bucket,
 } = require('@cumulus/aws-client/S3');
 const { randomString } = require('@cumulus/common/test-utils');
+
+const {
+  localStackConnectionEnv,
+  generateLocalTestDb,
+  destroyLocalTestDb,
+  CollectionPgModel,
+  fakeCollectionRecordFactory,
+} = require('@cumulus/db');
+
+const { migrationDir } = require('../../../../../lambdas/db-migration');
 const { createFakeJwtAuthToken, setAuthorizedOAuthUsers } = require('../../../lib/testUtils');
-
 const models = require('../../../models');
-
 const { app } = require('../../../app');
 
 process.env.AccessTokensTable = randomString();
 process.env.backgroundQueueUrl = randomString();
-process.env.CollectionsTable = randomString();
 process.env.GranulesTable = randomString();
 process.env.TOKEN_SECRET = randomString();
 
 const fakeCollectionId = 'FakeCollection___006';
-const fakeCollection = { duplicateHandling: 'replace' };
-
 let accessTokenModel;
 let jwtAuthToken;
 
-test.before(async () => {
+const testDbName = randomString(12);
+
+test.before(async (t) => {
+  process.env = {
+    ...process.env,
+    ...localStackConnectionEnv,
+    PG_DATABASE: testDbName,
+  };
+  const { knex, knexAdmin } = await generateLocalTestDb(testDbName, migrationDir);
+  t.context.knex = knex;
+  t.context.knexAdmin = knexAdmin;
+  const collectionPgModel = new CollectionPgModel();
+
+  await collectionPgModel.create(t.context.knex, fakeCollectionRecordFactory({ name: 'FakeCollection', version: '006' }));
+
   process.env.system_bucket = randomString();
   await s3().createBucket({ Bucket: process.env.system_bucket }).promise();
 
@@ -40,9 +59,15 @@ test.before(async () => {
   jwtAuthToken = await createFakeJwtAuthToken({ accessTokenModel, username });
 });
 
-test.after.always(async () => {
+test.after.always(async (t) => {
   await recursivelyDeleteS3Bucket(process.env.system_bucket);
   await accessTokenModel.deleteTable();
+
+  await destroyLocalTestDb({
+    knex: t.context.knex,
+    knexAdmin: t.context.knexAdmin,
+    testDbName,
+  });
 });
 
 test('put request with reingest action calls the granuleModel.reingest function with expected parameters', async (t) => {
@@ -51,10 +76,6 @@ test('put request with reingest action calls the granuleModel.reingest function 
   );
   const granuleReingestStub = sinon.stub(models.Granule.prototype, 'reingest').returns(
     new Promise((resolve) => resolve({ response: 'fakeResponse' }))
-  );
-
-  const collectionGetStub = sinon.stub(models.Collection.prototype, 'get').returns(
-    new Promise((resolve) => resolve(fakeCollection))
   );
 
   const body = {
@@ -76,5 +97,4 @@ test('put request with reingest action calls the granuleModel.reingest function 
 
   granuleGetStub.restore();
   granuleReingestStub.restore();
-  collectionGetStub.restore();
 });